ISSUE #4601 - move reusable colorPicker helpers to color helpers
