+++ frontend/src/v5/helpers/colors.helper.ts
@@ -16,6 +16,15 @@
  */
 
 import { contrastColor } from 'contrast-color';
+import { rgbToHex as muiRgbToHex, hexToRgb as muiHexToRgb } from '@mui/material';
+
+type GroupColor = {
+	color?: any,
+	opacity?: number,
+};
+export type RgbArray = [number, number, number];
+export type RgbGroupColor = GroupColor & { color?: RgbArray };
+export type HexGroupColor = GroupColor & { color?: string };
 
 export const hexToOpacity = (hex: string, opacityInPercentage: number): string => {
 	const formattedOpacity = Math.floor((opacityInPercentage / 100) * 255)
@@ -24,3 +33,20 @@ export const hexToOpacity = (hex: string, opacityInPercentage: number): string =
 };
 
 export const isLight = (hex: string, freshold?: number) => contrastColor({ bgColor: hex, threshold: freshold ?? 127 }) !== '#FFFFFF';
+
+export const rgbToHex = (color: RgbArray) => color && muiRgbToHex(`rgb(${color.join()})`);
+export const rgbGroupColorToHex = ({ opacity, color }: RgbGroupColor): HexGroupColor => ({
+	opacity,
+	color: rgbToHex(color),
+});
+
+const hexToRgb = (color) => color && muiHexToRgb(color).slice(4, -1).split(',').map(Number) as RgbArray;
+export const hexGroupColorToRgb = ({ opacity, color }: HexGroupColor): RgbGroupColor => ({
+	opacity,
+	color: hexToRgb(color),
+});
+
+export const hexToDecimal = (hex) => parseInt(hex, 16);
+export const decimalToHex = (decimal) => decimal.toString(16).padStart(2, '0');
+
+export const getColorHexIsValid = (color: string) => !color || /^#[0-9a-f]{6}/i.test(color);
