+++ frontend/src/v5/ui/controls/inputs/colorPicker/colorPicker.helpers.ts
@@ -15,32 +15,7 @@
  *  along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-import { rgbToHex as muiRgbToHex, hexToRgb as muiHexToRgb } from '@mui/material';
-
-type GroupColor = {
-	color?: any,
-	opacity?: number,
-};
-export type RgbArray = [number, number, number];
-export type RgbGroupColor = GroupColor & { color?: RgbArray };
-export type HexGroupColor = GroupColor & { color?: string };
-
-export const rgbToHex = (color: RgbArray) => color && muiRgbToHex(`rgb(${color.join()})`);
-export const rgbGroupColorToHex = ({ opacity, color }: RgbGroupColor): HexGroupColor => ({
-	opacity,
-	color: rgbToHex(color),
-});
-
-const hexToRgb = (color) => color && muiHexToRgb(color).slice(4, -1).split(',').map(Number) as RgbArray;
-export const hexGroupColorToRgb = ({ opacity, color }: HexGroupColor): RgbGroupColor => ({
-	opacity,
-	color: hexToRgb(color),
-});
-
-export const hexToDecimal = (hex) => parseInt(hex, 16);
-export const decimalToHex = (decimal) => decimal.toString(16).padStart(2, '0');
-
-export const getColorIsValid = (color: string) => !color || /[0-9a-f]{6}/.test(color?.toLocaleLowerCase());
+import { RgbArray } from '@/v5/helpers/colors.helper';
 
 export const UNSET_RGB_COLOR: RgbArray = [120, 120, 120];
 export const UNSET_HEX_COLOR = '#787878';
